requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
music-tag==0.4.3
mutagen==1.47.0
pillow==12.0.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
requests==2.32.5
shiboken6==6.10.0
urllib3==2.5.0
